requirements.txt
ddt==1.4.4
PyYAML==6.0.1
selenium==4.21.0
